Update bistro, PELUX and BSP layers
This introduces PELUX as a Yocto distribution. The distribution conf file
as well as the image files can be found in the meta-pelux layer. Distro
and image configuration has been moved from the BSP meta layers
local.conf files to the meta-pelux layer. The meta-bistro layer is used
as an aggregate of Pelagicore components and bbappends.

Signed-off-by: Gordan Markuš <gordan.markus@pelagicore.com>