Add archive step to Jenkinsfile
Before the artifacts never have been archived. This patch adds an
archive step to archive the images every time and to only archive
sstate-cache and downloads when we build the master branch.

We only archive sstate-cache and downloads on the master branch
build because it takes quite some time to copy all those artifacts
so we only want to do that during the nightly builds, but not
during the pull requests builds.