Do not rely on default values for arguments in do_repo_init
The default value of the second argument is "/vagrant". The default value
of the third argument is "pelux_user".

Third argument is changed to /home/yoctouser/pelux_user now. It worked as
expected before because pwd when running command in container was
/home/yoctouser . Better to be explicit.

Will change value of yoctoDir in following commit.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>