Remove code duplication in buildWithLayer()
By adding layer arguments to buildManifest() and replace layer in
buildManifest() if they are both set to something other than "".

When env.SMOKE_TEST != "true" and env.NIGHTLY_BUILD != "true",
buildManifest() is exactly the same as the old buildWithLayer().

Keep buildWithLayer() as a thin wrapper around buildManifest() since it is
called from Jenkinsfile:s in other repositories. Possible to remove in the
future if we modify all repositories to use buildManifest() directly
instead.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>