Archive: remove the downloads folder from archive
Storing the downloads folder is unnecessary for us since we store the
yocto cache already. The downloads folders also takes up 10-20GB per
variant and since we want to minimise the amount of memory each build takes
(to be able to store more of them without increasing storage capacity) we
will remove it now.

Signed-off-by: Therese Nordqvist <therese.nordqvist@pelagicore.com>