test-scripts/local.conf: remove auto tests temporarily
Test scripts have been failing because of a package (galculator) being unable
to be tar-ed. Its format is bz2 but neither bzip2, or lbzip2 was installed
on the target. This is not a fix, this is a temporarily workaround.

Signed-off-by: Fisnik Hajredini <fhajredini@luxoft.com>
(cherry picked from commit 15daff10bcd90d01757aac3e9045b390bd1c01b5)