Do not build -dev images automatically
When you build the -dev image automatically, it means that
when you try to actually build a dev image, it will go on the
update image as well and do the "-dev-dev-update". This is
inconsistent with the other builds that we do and whenever we try to
build rpi -dev images, it always fails because of not finding the
specified image.

Signed-off-by: Fisnik Hajredini <fhajredini@luxoft.com>