copy-to-archive: copy only the timestamped image
Currently we copy duplicate images. One with the timestamp and the other
without the timestamp. This is due to there being a symlink in the
tmp/deploy/images folder. E.g.:

core-image-pelux-qtauto-neptune-dev-intel-corei7-64-20190307122631.rootfs.wic
core-image-pelux-qtauto-neptune-dev-intel-corei7-64.wic ->
core-image-pelux-qtauto-neptune-dev-intel-corei7-64-20190307122631.rootfs.wic

When archiving we actually copy what the symlinks points to so the same
file is copied twice.

This change copies only the one with a timestamp.
Signed-off-by: Fisnik Hajredini <fisnikhajredini@gmail.com>