yocto.groovy: make yocto compatibility tests optional
Yocto compatibility tests produce inaccurate and confusing results, e.g
they looks for dependencies of every existing package in the tested
layer and if it is not found, the test fails, even though the dependency
is present in another layer that is not used by PELUX and therefore not
included in our bblayers.conf.

Tests also do not distinguish between critical and non-critical errors.

Those tests should occasionally be executed locally instead to ensure
PELUX layers are compatible with the Yocto project.

Added flag that allows enabling/disabling those tests through Jenkins
parameters, which is set to false (disabled) by default.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>