yocto.groovy: remove smoketest build conf setup
Smoketest configuration was appended to the qemu build
local.conf during the smoketest job running. This conf
is now part of the qemu variant local.conf, so there is
no need to append it to the build conf during the jenkins
test job. This commit removes the code that updated the
qemu build local.conf file.

Signed-off-by: Tariq Ansari <tansari@luxoft.com>