yocto.groovy: add parameter for accepting Freescale EULA
To be able to build some of the Freescale BSP packages, one must accept
the Freescale EULA by adding the following line to local.conf:

  ACCEPT_FSL_EULA = "1"

This line is commented by default to not force a user who builds PELUX
to accept the license without their consent.

Add Jenkins parameter ACCEPT_FSL_EULA that is set to false by default,
which in the case is set to true will tell the groovy script to
uncomment the line in local.conf and accept the license agreement.

Signed-off-by: Oleksandr Kravchuk <oleksandr.kravchuk@pelagicore.com>